+++ Sourcery/Models/Method.swift
@@ -87,6 +87,9 @@ final class Method: NSObject, AutoDiffable, Annotated, NSCoding {
         return returnTypeName.unwrappedTypeName
     }
 
+    /// Whether this method throws or rethrows
+    let `throws`: Bool
+
     /// Method access level
     let accessLevel: String
 
@@ -116,6 +119,7 @@ final class Method: NSObject, AutoDiffable, Annotated, NSCoding {
     init(selectorName: String,
          parameters: [MethodParameter] = [],
          returnTypeName: TypeName = TypeName("Void"),
+         throws: Bool = false,
          accessLevel: AccessLevel = .internal,
          isStatic: Bool = false,
          isClass: Bool = false,
@@ -126,6 +130,7 @@ final class Method: NSObject, AutoDiffable, Annotated, NSCoding {
         self.selectorName = selectorName
         self.parameters = parameters
         self.returnTypeName = returnTypeName
+        self.throws = `throws`
         self.accessLevel = accessLevel.rawValue
         self.isStatic = isStatic
         self.isClass = isClass
@@ -140,6 +145,7 @@ final class Method: NSObject, AutoDiffable, Annotated, NSCoding {
             guard let parameters: [MethodParameter] = aDecoder.decode(forKey: "parameters") else { NSException.raise(NSExceptionName.parseErrorException, format: "Key '%@' not found.", arguments: getVaList(["parameters"])); fatalError() }; self.parameters = parameters
             guard let returnTypeName: TypeName = aDecoder.decode(forKey: "returnTypeName") else { NSException.raise(NSExceptionName.parseErrorException, format: "Key '%@' not found.", arguments: getVaList(["returnTypeName"])); fatalError() }; self.returnTypeName = returnTypeName
             self.returnType = aDecoder.decode(forKey: "returnType")
+            self.`throws` = aDecoder.decode(forKey: "throws")
             guard let accessLevel: String = aDecoder.decode(forKey: "accessLevel") else { NSException.raise(NSExceptionName.parseErrorException, format: "Key '%@' not found.", arguments: getVaList(["accessLevel"])); fatalError() }; self.accessLevel = accessLevel
             self.isStatic = aDecoder.decode(forKey: "isStatic")
             self.isClass = aDecoder.decode(forKey: "isClass")
@@ -155,6 +161,7 @@ final class Method: NSObject, AutoDiffable, Annotated, NSCoding {
             aCoder.encode(self.parameters, forKey: "parameters")
             aCoder.encode(self.returnTypeName, forKey: "returnTypeName")
             aCoder.encode(self.returnType, forKey: "returnType")
+            aCoder.encode(self.`throws`, forKey: "throws")
             aCoder.encode(self.accessLevel, forKey: "accessLevel")
             aCoder.encode(self.isStatic, forKey: "isStatic")
             aCoder.encode(self.isClass, forKey: "isClass")