+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Master
 
 ### New Features
+- Added support for method return types with `throws` and `rethrows`
 - Added a new filter `replace`. Usage: `{{ name|replace:"substring","replacement" }}` - replaces occurrences of `substring` with `replacement` in `name` (case sensitive)
 - Sourcery will now use parallel parsing, expect more than 2x as fast execution.
 - Sourcery will now cache source artifacts, in many scenarios it will lead to order of magnitude faster processing.